kinda fix department part

+++ src/main/java/com/collage/project/Department/controller/DepartmentController.java
@@ -3,6 +3,8 @@ package com.collage.project.Department.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -33,8 +35,6 @@ public class DepartmentController {
         if(deptName != null){
             deptName = String.format("%%%s%%", deptName); // 格式化用于 like 搜索
             return departmentService.getDepartmentsByName(deptName);
-        }else if(name != null){ // 精确查询
-            return departmentService.getDepartmentsBySpecificName(name);
         }else{
             return departmentService.getDepartments();
         }
@@ -48,19 +48,33 @@ public class DepartmentController {
 
     // 添加部门
     @PostMapping
-    public void addDepartment(@RequestBody Department department) {
-        departmentService.addDepartment(department);
+    public ResponseEntity<String> addDepartment(@RequestBody Department department) {
+        // 判断部门重名
+        if(departmentService.getDepartmentsCountByName(department.getDeptName()) == 0){
+            departmentService.addDepartment(department);
+            return new ResponseEntity<>("success", HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>("failed", HttpStatus.CONFLICT);
     }
 
     // 更新部门
     @PutMapping(path = "{id}")
-    public void updateDepartment(@RequestBody Department department, @PathVariable("id") int deptId){
+    public ResponseEntity<String> updateDepartment(@RequestBody Department department, @PathVariable("id") int deptId){
+        // 判断部门重名
+        if(departmentService.getDepartmentsCountByName(department.getDeptName()) == 0){
         departmentService.updateDepartment(department);
+            return new ResponseEntity<>("success", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("failed", HttpStatus.CONFLICT);
     }
 
     // 删除部门
     @DeleteMapping(path = "{id}")
-    public void deleteDepartment(@PathVariable("id") int deptId){
-        departmentService.deleteDepartment(deptId);
+    public ResponseEntity<String> deleteDepartment(@PathVariable("id") int deptId){
+        if(departmentService.getRelatedEmployeesCountById(deptId) == 0){
+            departmentService.deleteDepartment(deptId);
+            return new ResponseEntity<>("success", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("failed", HttpStatus.CONFLICT);
     }
 }
